requirements.txt
Django>=5.1
djangorestframework>=3.15.2
django-model-utils>=4.5.1
Faker>=27.0.0
